:construction_worker: Runtime test the app from the CI
Leverage `reactivecircus/android-emulator-runner` to run the testapps
artifact on an emulator and check the ADB logs.
We're currently checking for the following pattern from the logcat output:
- "I python: Initialized python"
- "I python: Ran 14 tests in"
- "I python: OK"

Note that we're keeping things simple for the first iteration and we only
run the `ubuntu-latest-sdl2-artifacts` artifact.
In the future we may want to test the different bootstraps, build host
and even updated recipes dynamically.

Also bump actions/checkout to v5.